Split test_report.py into 4 focused files (issue #495)
- Move _close_sqlite_datalayers autouse fixture to conftest.py
- test_create_report_received.py: CreateReport creation, no-standalone-status, duplicate handling (3 classes, 7 tests)
- test_submit_report_received.py: SubmitReport log messages, case creation, offer addressing (3 classes, 11 tests)
- test_report_case_level.py: InvalidateReport, CloseReport, and delegation to case use cases (2 classes, 8 tests)
- test_ack_validate_report_received.py: AckReport, ValidateReport, and full-flow integration (3 classes, 10 tests)
- Remove test_report.py (1154 lines → 4 files, each ≤400 lines)

All 36 tests pass. Black, flake8, mypy, pyright clean.
